+++ database/install/myems_historical_db.sql
@@ -18,6 +18,7 @@ CREATE TABLE IF NOT EXISTS `myems_historical_db`.`tbl_analog_value` (
   `utc_date_time` DATETIME NOT NULL,
   `actual_value` DECIMAL(21, 6) NOT NULL,
   `is_bad` BOOL,
+  `is_published` BOOL,
   PRIMARY KEY (`id`));
 CREATE INDEX `tbl_analog_value_index_1` ON `myems_historical_db`.`tbl_analog_value` (`point_id`, `utc_date_time`);
 CREATE INDEX `tbl_analog_value_index_2` ON `myems_historical_db`.`tbl_analog_value` (`utc_date_time`);
@@ -62,6 +63,7 @@ CREATE TABLE IF NOT EXISTS `myems_historical_db`.`tbl_digital_value` (
   `utc_date_time` DATETIME NOT NULL,
   `actual_value` INT NOT NULL,
   `is_bad` BOOL,
+  `is_published` BOOL,
   PRIMARY KEY (`id`));
 CREATE INDEX `tbl_digital_value_index_1` ON `myems_historical_db`.`tbl_digital_value` (`point_id`, `utc_date_time`);
 CREATE INDEX `tbl_digital_value_index_2` ON `myems_historical_db`.`tbl_digital_value` (`utc_date_time`);
@@ -92,6 +94,7 @@ CREATE TABLE IF NOT EXISTS `myems_historical_db`.`tbl_energy_value` (
   `utc_date_time` DATETIME NOT NULL,
   `actual_value` DECIMAL(21, 6) NOT NULL,
   `is_bad` BOOL,
+  `is_published` BOOL,
   PRIMARY KEY (`id`));
 CREATE INDEX `tbl_energy_value_index_1` ON `myems_historical_db`.`tbl_energy_value` (`point_id`, `utc_date_time`);
 CREATE INDEX `tbl_energy_value_index_2` ON `myems_historical_db`.`tbl_energy_value` (`utc_date_time`);
@@ -123,6 +126,7 @@ CREATE TABLE IF NOT EXISTS `myems_historical_db`.`tbl_text_value` (
   `utc_date_time` DATETIME NOT NULL,
   `actual_value` LONGTEXT NOT NULL,
   `is_bad` BOOL,
+  `is_published` BOOL,
   PRIMARY KEY (`id`));
 CREATE INDEX `tbl_text_value_index_1` ON `myems_historical_db`.`tbl_text_value` (`point_id`, `utc_date_time`);
 CREATE INDEX `tbl_text_value_index_2` ON `myems_historical_db`.`tbl_text_value` (`utc_date_time`);