+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added control modes and control times
 - added legend to MultipleLineChart in myems-web
 - added power integrator table in database
+- added columns is_published to historical database
 ### Changed
 - replaced datetime.strftime() with datetime.isoformat() in myems-api
 - changed space reports to automaticly submit in myems-web